relay/requirements.txt
rpi.gpio
paho-mqtt
asyncio-mqtt
